refactor: split sections into components
